creates GenericImageBuilderTestSuite
for use by multiple cloud implementations
